refactoring: rename parser file to schema_parser
