+++ xsd-parser-rs/src/parser/complex_type.rs
@@ -2,7 +2,7 @@ use std::cell::RefCell;
 
 use roxmltree::Node;
 
-use crate::parser::parser::parse_node;
+use crate::parser::schema_parser::parse_node;
 use crate::parser::types::{RsEntity, Struct, StructField, StructFieldSource};
 use crate::parser::utils::{
     any_attribute_field, attributes_to_fields, find_child, get_documentation, get_parent_name,
