+++ xsd-parser-rs/src/parser/element.rs
@@ -1,7 +1,7 @@
 use roxmltree::Node;
 
 use crate::parser::constants::attribute;
-use crate::parser::parser::parse_node;
+use crate::parser::schema_parser::parse_node;
 use crate::parser::types::{
     Alias, EnumCase, RsEntity, StructField, StructFieldSource, TypeModifier,
 };
