+++ xsd-parser-rs/src/parser/mod.rs
@@ -5,7 +5,7 @@ mod complex_type;
 pub mod constants;
 mod element;
 mod list;
-mod parser;
+mod schema_parser;
 mod sequence;
 mod simple_content;
 mod simple_type;
@@ -13,4 +13,4 @@ mod tests;
 pub mod types;
 mod utils;
 pub mod xsd_elements;
-pub use parser::*;
+pub use schema_parser::*;
