+++ xsd-parser-rs/src/parser/schema_parser.rs
